+++ src/db/delivery/query/packing_list_entry.js
@@ -95,8 +95,8 @@ export async function selectAll(req, res, next) {
 			concat(oe.style, ' / ', oe.color, ' / ', oe.size) as style_color_size,
 			oe.quantity::float8 as order_quantity,
 			sfg.uuid as sfg_uuid,
-			sfg.warehouse as warehouse,
-			sfg.delivered as delivered,
+			sfg.warehouse::float8 as warehouse,
+			sfg.delivered::float8 as delivered,
 			(oe.quantity - sfg.warehouse)::float8 as balance_quantity
 		FROM 
 			delivery.packing_list_entry ple
